+++ pkg/breakglass/debug/debug_session_reconciler_workload.go
@@ -747,7 +747,7 @@ func resourceMayBeDeletedByDebugSession(obj *unstructured.Unstructured, ds *brea
 	case hasDebugSessionLegacyMarker(obj):
 		return resourceOwnedByLegacyDebugSession(obj, ds)
 	default:
-		return true
+		return false
 	}
 }